requirements.txt
fastapi
uvicorn
pymupdf
python-multipart
Pillow
opencv-python
numpy

